% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_args}
\alias{check_args}
\title{Capture Arguments}
\usage{
check_args(...)
}
\arguments{
\item{...}{unquoted arguments names}
}
\value{
List of quosures.
}
\description{
Helper to catch arguments.
}
\examples{
\dontrun{
check_args(x, y)
}
}
\keyword{internal}
