% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_length}
\alias{check_length}
\title{Check if Argument is of Proper Length}
\usage{
check_length(x, arg_length = 1L, alt_null = FALSE, add_msg = NULL)
}
\arguments{
\item{x}{Function arguments that are being asserted.}

\item{arg_length}{Integer. Length of argument, for scalars it should take
value \code{1} (default).}

\item{alt_null}{Logical. Should argument accept NULL value.}

\item{add_msg}{Is an additional message that can be printed over the standard
function error message. You can:
\itemize{
\item pass the names of the arguments that failed the test by using
\code{{x_name}} in the message body (e.g. "What are the \{wrong_names\}");
\item pass the tested length by using \code{{arg_length}} in the message body (e.g.
"I want them to be \{arg_length\})"
\item pass the lengths of the arguments that failed the test by using
\code{{wrong_length}} in the message body (e.g. "\{wrong_lengths\} are wrong")
}}
}
\value{
Returns invisible \code{NULL} when argument is of asserted length,
otherwise it will throw an error.
}
\description{
TODO.
}
\keyword{internal}
