% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_null_cond}
\alias{check_null_cond}
\title{Return check_null Value}
\usage{
check_null_cond(x, alt_null)
}
\arguments{
\item{x}{Argument to check if is NULL.}

\item{alt_null}{Logical. If \code{TRUE} it will check if \code{x} is \code{NULL}.}
}
\value{
Either \code{TRUE} or \code{FALSE}.
}
\description{
Check if \code{alt_null} argument is TRUE or FALSE. If it is \code{FALSE}
it will return \code{FALSE}. If the argument is \code{TRUE} it will check if the
x argument is \code{NULL} and return logical value.
}
\keyword{internal}
