% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_setup.R
\name{cloud_drive_attach}
\alias{cloud_drive_attach}
\title{Attach Google Drive folder to project}
\usage{
cloud_drive_attach(project = ".")
}
\arguments{
\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\value{
This function does not return a meaningful value. Its primary purpose
is the side effect of updating the project's DESCRIPTION file with the
associated Google Drive folder identifier.
}
\description{
This function facilitates the association of a specific Google
Drive folder with a project by adding a unique identifier to the project's
DESCRIPTION file. The user is prompted to navigate to the Google Drive
website, select or create the desired folder for the project, and then
provide its URL. The function extracts the necessary information from the
URL and updates the \code{cloudfs.drive} field in the DESCRIPTION file
accordingly.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cloud_drive_attach()
\dontshow{\}) # examplesIf}
}
