% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{cloud_guess_read_fun}
\alias{cloud_guess_read_fun}
\title{Guess reading function based on file extensions}
\usage{
cloud_guess_read_fun(file)
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}
}
\description{
Take a look at the switch call. That's basically it. Returns an
appropriate function or throws an error if wasn't able to find one.
}
\section{Default reading functions}{


Here's how we identify a reading function based on file extension
\itemize{
\item \code{.csv}: \link[readr:read_delim]{readr::read_csv}
\item \code{.json}: \link[jsonlite:read_json]{jsonlite::read_json}
\item \code{.rds}: \link[base:readRDS]{base::readRDS}
\item \code{.sav}: \link[haven:read_spss]{haven::read_sav}
\item \code{.xls}: \link{cloud_read_excel}
\item \code{.xlsx}: \link{cloud_read_excel}
\item \code{.xml}: \link[xml2:read_xml]{xml2::read_xml}
}
}

\keyword{internal}
