% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{cloud_guess_write_fun}
\alias{cloud_guess_write_fun}
\title{Guess writing function based on file extensions}
\usage{
cloud_guess_write_fun(file)
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}
}
\description{
Take a look at the switch call. That's basically it. Returns an
appropriate function or throws an error if wasn't able to find one.
}
\section{Default writing functions}{


Here's how we identify a writing function based on file extension
\itemize{
\item \code{.csv}: \link[readr:write_delim]{readr::write_csv}
\item \code{.json}: \link[jsonlite:read_json]{jsonlite::write_json}
\item \code{.rds}: \link[base:readRDS]{base::saveRDS}
\item \code{.xls}: \link[writexl:write_xlsx]{writexl::write_xlsx}
\item \code{.xlsx}: \link[writexl:write_xlsx]{writexl::write_xlsx}
\item \code{.sav}: \link[haven:read_spss]{haven::write_sav}
\item \code{.xml}: \link[xml2:write_xml]{xml2::write_xml}
}
}

\keyword{internal}
