% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{cloud_not_wd_warning}
\alias{cloud_not_wd_warning}
\title{Warn if cloud function is used not for current working directory}
\usage{
cloud_not_wd_warning(project)
}
\arguments{
\item{project}{Path to a project. By default it is current working directory.}
}
\description{
Functions for uploading/downloading files from project cloud
locations are designed to synchronize local and cloud folder structures.
That is e.g. when you call \code{cloud_s3_upload} with \code{file} parameter set to
"data/demo.csv" and \code{project} parameter set to something different from the
current working directory it is always assumed that "data/demo.csv" from
\strong{the project's folder} and not from the current working directory needs
to be uploaded to S3. But for development purposes it is handy to be able
to call the functions not only for the current wd. This function checks
that project is set to the current wd. If not, it throws a warning and asks
if user wants to continue.
}
