% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_local.R
\name{cloud_object_prep_bulk}
\alias{cloud_object_prep_bulk}
\title{Prepare object content dataframe to be used by bulk download/read
functions}
\usage{
cloud_object_prep_bulk(content, quiet = FALSE)
}
\arguments{
\item{content}{(data.frame) output of \code{cloud_object_ls()}}

\item{quiet}{all caution messages may be turned off by setting this parameter
to \code{TRUE}.}
}
\description{
\code{cloud_object_ls} returns an ls-like dataframe for a named list
of objects. This function is used mainly to inform the user about which
files are going to be written and to ask for confirmation
}
\keyword{internal}
