% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write.R
\name{cloud_read_excel}
\alias{cloud_read_excel}
\title{Read excel file as a list of dataframes}
\usage{
cloud_read_excel(path)
}
\arguments{
\item{path}{Path to the xlsx/xls file.}
}
\value{
A named list of dataframes, where each dataframe corresponds to a
sheet in the Excel file. The names of the list elements are derived from
the sheet names.
}
\description{
Uses \link[readxl:read_excel]{readxl::read_excel} under the hood, reads all sheets and
returns them as a named list of dataframes.
}
\examples{
datasets <- readxl::readxl_example("datasets.xlsx")
cloud_read_excel(datasets)

}
