% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_setup.R
\name{cloud_s3_attach}
\alias{cloud_s3_attach}
\title{Attach S3 folder to project}
\usage{
cloud_s3_attach(project = getwd())
}
\arguments{
\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\description{
The \code{cloud_s3_attach()} function is used to add a field to the
DESCRIPTION file of a project, which uniquely identifies the location of
the project's folder in an S3 cloud storage. It prompts the user to visit
the S3 cloud storage website (https://s3.console.aws.amazon.com/) where
they can find or create a dedicated folder for the project. After the user
has selected or created the desired folder, they can copy the URL of the
folder from the web browser and paste it into the R console. The function
then parses the URL and populates the corresponding field (CloudS3) in the
DESCRIPTION file with a string that represents the location of the project
in the S3 cloud storage.
}
\examples{
\dontrun{cloud_s3_attach()}

}
