% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{cloud_validate_file_path}
\alias{cloud_validate_file_path}
\title{Validate file path for cloud functions}
\usage{
cloud_validate_file_path(file, error = TRUE)
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}

\item{error}{if \code{TRUE} (default), throws an error if \code{file} is not a valid
file path.}
}
\description{
Makes sure that file path passed to a cloud function is in the
right format.
}
