% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc_file}
\alias{doc_file}
\title{Package-wide description of \code{file} parameter}
\usage{
doc_file(file)
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}
}
\description{
A dummy function to be referred by \verb{@inheritParams} for a
parameter documentation.
}
\keyword{internal}
