% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc.R
\name{doc_local}
\alias{doc_local}
\title{Package-wide description of \code{local} parameter}
\usage{
doc_local(local)
}
\arguments{
\item{local}{Logical, defaulting to \code{FALSE}. If \code{TRUE}, the function will
also create a local copy of the file at the specified path. Note that some
writing functions might not overwrite existing files unless explicitly
allowed. Typically, such functions have a parameter (often named
\code{overwrite}) to control this behavior. Check the documentation of the
writing function used to determine the exact parameter name and pass it
through the \code{...} argument if necessary. Alternatively, you can define an
anonymous function for \code{fun} that calls a writing function with the
overwriting option enabled.}
}
\description{
A dummy function to be referred by \verb{@inheritParams} for a
parameter documentation.
}
\keyword{internal}
