% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{proj_desc_get}
\alias{proj_desc_get}
\title{Extract values from DESCRUPTION file}
\usage{
proj_desc_get(key, project = ".")
}
\arguments{
\item{key}{Character. What field to search for in DESCRIPTION file.}

\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\description{
Extract values from DESCRUPTION file
}
