% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{validate_desc}
\alias{validate_desc}
\title{Validate project's DESCRIPTION file}
\usage{
validate_desc(project = ".")
}
\arguments{
\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\description{
Checks that DESCRIPTION file exists in a project folder. If it's
not the case, proposes to create a DESCRIPTION file from template.
}
\keyword{internal}
