% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{check_scalar}
\alias{check_scalar}
\title{Check if Function Argument is Scalar}
\usage{
check_scalar(..., arg_class, alt_null = FALSE)
}
\arguments{
\item{...}{Function argument that is being asserted.}

\item{arg_class}{Class name. Usually "character", "numeric",
"data.frame", etc.}

\item{alt_null}{Logical. Should argument accept NULL value.}
}
\value{
Invisible \code{NULL} if assertion is \code{TRUE}, otherwise an error message.
}
\description{
A function to check that argument is of proper class and of
length 1.
}
\keyword{internal}
