% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_setup.R
\name{cloud_drive_attach}
\alias{cloud_drive_attach}
\title{Attach Google Drive folder to project}
\usage{
cloud_drive_attach(project = ".")
}
\arguments{
\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\description{
The \code{cloud_drive_attach()} function is designed to add a field
to the DESCRIPTION file of a project, which uniquely identifies the
location of the project's folder in Google Drive. The function prompts the
user to visit the Google Drive website (https://drive.google.com/) where
they can find or create a dedicated folder for the project. Once the user
has located or created the desired folder, they can copy the URL of the
folder from the web browser and paste it into the R console. The function
then parses the URL and populates the corresponding field (cloudfs.drive)
in the DESCRIPTION file with a string that represents the location of the
project in Google Drive.
}
\examples{
\dontrun{cloud_drive_attach()}

}
