% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_setup.R
\name{cloud_drive_attach}
\alias{cloud_drive_attach}
\title{Attach Google Drive folder to project}
\usage{
cloud_drive_attach(project = getwd())
}
\arguments{
\item{project}{Path to a project. By default it is current working directory.}
}
\description{
Launches a dialogue asking user to provide a link to a Google
Drive to be attached as the dedicated Google Drive folder of the project.
}
\examples{
\dontrun{cloud_drive_attach()}

}
