% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_show.R
\name{cloud_drive_browse}
\alias{cloud_drive_browse}
\title{Browse project's Google Drive folder}
\usage{
cloud_drive_browse(path = "", project = getwd())
}
\arguments{
\item{path}{(optional) Path inside Google Drive folder to open. By default,
when \code{path = ""}, navigates to the root level of project's folder.}

\item{project}{Path to a project. By default it is current working directory.}
}
\description{
Opens project's Google Drive folder in browser.
}
\details{
Google Drive file structure is different from the usual file
structure like e.g. on Linux or Windows. A folder on Google Drive can have
two or more child folders with the same name. Google Drive marks files and
folders with so-called id values to distinguish between them. These values
are always unique. You can see them in browser URL for example. The concept
of "name" is in the first place for convenience of the end user.

In such a setup a relative file path may correspond to multiple files or
folders. This function however works under assumption that the relative
path you pass to it defines strictly one object. If there's any ambiguity
it throws an error.
}
\examples{
\dontrun{
cloud_drive_browse()
cloud_drive_browse("models/kmeans")
}

}
