% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_transfer_bulk.R
\name{cloud_drive_download_bulk}
\alias{cloud_drive_download_bulk}
\title{Download Google Drive contents in bulk}
\usage{
cloud_drive_download_bulk(content, quiet = FALSE)
}
\arguments{
\item{content}{(data.frame) Output of \code{cloud_drive_ls()}}

\item{quiet}{All caution messages may be turned off by setting this parameter
to \code{TRUE}.}
}
\description{
\link{cloud_drive_ls} function returns a dataframe of contents of a
Google Drive folder. \code{cloud_drive_download_bulk} can be applied to such a
dataframe to download all the listed files.

The workflow in mind is that you would call \code{cloud_drive_ls()}, then use
dplyr verbs to keep only files that you need and then call
\code{cloud_drive_download_bulk} on the result.
}
\examples{
\dontrun{
cloud_drive_ls("data") \%>\% 
  filter(type == "csv") \%>\% 
  cloud_drive_download_bulk()
}
  
}
