% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_transfer.R
\name{cloud_drive_upload}
\alias{cloud_drive_upload}
\title{Upload a local file to Google Drive}
\usage{
cloud_drive_upload(file, project = getwd())
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}

\item{project}{Path to a project. By default it is current working directory.}
}
\description{
Uploads a file from project's folder to the corresponding
location on project's Google Drive folder
}
\details{
Google Drive file structure is different from the usual file
structure like e.g. on Linux or Windows. A folder on Google Drive can have
two or more child folders with the same name. Google Drive marks files and
folders with so-called id values to distinguish between them. These values
are always unique. You can see them in browser URL for example. The concept
of "name" is in the first place for convenience of the end user.

In such a setup a relative file path may correspond to multiple files or
folders. This function however works under assumption that the relative
path you pass to it defines strictly one object. If there's any ambiguity
it throws an error.
}
\examples{
\dontrun{
# uploads data/demo.csv to 'data' subfolder of project's Google Drive folder
cloud_drive_upload("data/demo.csv")
}

}
