% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drive_transfer_bulk.R
\name{cloud_drive_upload_bulk}
\alias{cloud_drive_upload_bulk}
\title{Upload files to Google Drive in bulk}
\usage{
cloud_drive_upload_bulk(content, quiet = FALSE, project = getwd())
}
\arguments{
\item{content}{(data.frame) Output of \code{cloud_s3_ls()}}

\item{quiet}{All caution messages may be turned off by setting this parameter
to \code{TRUE}.}
}
\description{
\link{cloud_local_ls} function returns a dataframe of contents of
local project folder. \code{cloud_drive_upload_bulk} can be applied to such a
dataframe to upload all the listed files.

The workflow in mind is that you would call \code{cloud_local_ls()}, then use
dplyr verbs to keep only files that you need and then call
\code{cloud_drive_upload_bulk} on the result.
}
\examples{
\dontrun{
cloud_local_ls("plots") \%>\% 
  filter(type == "png") \%>\% 
  cloud_drive_upload_bulk()
}
  
}
