% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{cloud_get_roots}
\alias{cloud_get_roots}
\title{Get cloud roots of a project}
\usage{
cloud_get_roots(project = ".")
}
\arguments{
\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\description{
Returns a list with all \verb{cloudfs.*} roots defined in a project's
DESCRIPTION.
}
