% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloud_local.R
\name{cloud_local_ls}
\alias{cloud_local_ls}
\title{List Contents of local project folder}
\usage{
cloud_local_ls(
  path = "",
  root = ".",
  recursive = FALSE,
  full_names = FALSE,
  ignore = TRUE
)
}
\arguments{
\item{path}{(optional) Path inside local project folder to list contents of
a subfolder. By default, when \code{path = ""}, lists root-level files and
folders.}

\item{root}{Local path relative to which to consider all paths.}

\item{recursive}{(logical) If \code{TRUE}, lists contents recursively in all
nested subfolders. Default is \code{FALSE}.}

\item{full_names}{(logical) If \code{TRUE}, folder path is appended to object
names to give a relative file path.}

\item{ignore}{(logical) Currently just ignores the "renv" folder if \code{TRUE}.
The main reason for this parameter is that "renv" folder usually contains
thousands of files and it takes a lot of time to calculate its size. But
potentially we may use something like global or project-level cloud ignore
files akin to .gitignore.}
}
\description{
Prints names, timestamps and sizes of files and folders inside
local project folder.
}
\examples{
\dontrun{
# list only root-level files and folders
cloud_local_ls() 

# list all files in all nested folders
cloud_local_ls(recursive = TRUE)

# list contents of "plots/barplots" subfolder
cloud_local_ls("plots/barplots")
}

}
