% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_setup.R
\name{cloud_s3_attach}
\alias{cloud_s3_attach}
\title{Attach S3 folder to project}
\usage{
cloud_s3_attach(project = ".")
}
\arguments{
\item{project}{Character. Path to a project. By default it is current working
directory.}
}
\value{
This function does not return a meaningful value but modifies the
DESCRIPTION file of the specified project to include the S3 folder path.
}
\description{
This function facilitates the association of a specific S3
folder with a project by adding a unique identifier to the project's
DESCRIPTION file. The user is prompted to navigate to the S3 console,
select or create the desired folder for the project, and then provide its
URL. The function extracts the necessary information from the URL and
updates the \code{cloudfs.s3} field in the DESCRIPTION file accordingly.
}
\examples{
\dontrun{cloud_s3_attach()}

}
