% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_show.R
\name{cloud_s3_browse}
\alias{cloud_s3_browse}
\title{Browse project's S3 folder}
\usage{
cloud_s3_browse(path = "", root = NULL)
}
\arguments{
\item{path}{(optional) Path inside S3 folder to open. By default, when
\code{path = ""}, navigates to the root level of project's S3 folder.}

\item{root}{S3 path of the project root -- point relative to which to
consider all relative paths. When left as \code{NULL}, the root is automatically
derived from the \code{cloudfs.s3} field of the project's DESCRIPTION file.}
}
\description{
Opens project's S3 folder in browser.
}
\examples{
\dontrun{
cloud_s3_browse()
cloud_s3_browse("data")
}

}
