% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_transfer.R
\name{cloud_s3_download}
\alias{cloud_s3_download}
\title{Download a file from S3 to local project folder}
\usage{
cloud_s3_download(file, root = NULL)
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}

\item{root}{S3 path of the project root. This serves as the reference point
for all relative paths. When left as \code{NULL}, the root is automatically
derived from the \code{cloudfs.s3} field of the project's DESCRIPTION file.}
}
\description{
Downloads a file from project's S3 folder to local project
folder and saves it preserving folder structure.
}
\examples{
\dontrun{
# downloads data/demo.csv from project's S3 folder
# and saves it to local 'data' folder
cloud_s3_download("data/demo.csv")
}

}
