% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_transfer_bulk.R
\name{cloud_s3_download_bulk}
\alias{cloud_s3_download_bulk}
\title{Download S3 contents in bulk}
\usage{
cloud_s3_download_bulk(content, quiet = FALSE, project = getwd())
}
\arguments{
\item{content}{(data.frame) Output of \code{cloud_s3_ls()}}

\item{quiet}{All caution messages may be turned off by setting this parameter
to \code{TRUE}.}
}
\description{
\link{cloud_s3_ls} function returns a dataframe of contents of an S3
folder. \code{cloud_s3_download_bulk} can be applied to such a dataframe to
download all the listed files. \link{cloud_s3_download} is used under the hood.

The workflow in mind is that you would call \code{cloud_s3_ls()}, then use dplyr
verbs to keep only files that you need and then call \code{cloud_s3_download_bulk}
on the result.
}
\examples{
\dontrun{
cloud_s3_ls("data") \%>\% 
  filter(type == "csv") \%>\% 
  cloud_s3_download_bulk()
}
  
}
