% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_show.R
\name{cloud_s3_ls}
\alias{cloud_s3_ls}
\title{List Contents of Project's S3 Folder}
\usage{
cloud_s3_ls(path = "", recursive = FALSE, full_names = FALSE, root = NULL)
}
\arguments{
\item{path}{(optional) Path inside S3 folder to list contents of a subfolder.
By default, when \code{path = ""}, lists root-level files and folders.}

\item{recursive}{(logical) If \code{TRUE}, lists contents recursively in all
nested subfolders. Default is \code{FALSE}.}

\item{full_names}{(logical) If \code{TRUE}, folder path is appended to object
names to give a relative file path.}

\item{root}{S3 path of the project root -- point relative to which to
consider all relative paths. When left as \code{NULL}, the root is automatically
derived from the \code{cloudfs.s3} field of the project's DESCRIPTION file.}
}
\description{
Prints names, timestamps and sizes of files and folders inside
S3 folder.
}
\examples{
\dontrun{
# list only root-level files and folders
cloud_s3_ls() 

# list all files in all nested folders
cloud_s3_ls(recursive = TRUE)

# list contents of "plots/barplots" subfolder
cloud_s3_ls("plots/barplots")
}

}
