% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_transfer.R
\name{cloud_s3_read}
\alias{cloud_s3_read}
\title{Read an object from S3}
\usage{
cloud_s3_read(file, fun = NULL, ..., project = getwd())
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}

\item{fun}{Reading function. By default is \code{NULL} which means that it will
be attempted to guess an appropriate reading function from file extension.}

\item{...}{Further parameters to pass to \code{fun}.}

\item{project}{Path to a project. By default it is current working directory.}
}
\description{
Reads a file from project's S3. This function tries to guess an
appropriate reading function based on the \code{file} name, but you can also
provide a function by yourself if it's needed.
}
\section{Default reading functions}{


Here's how we identify a reading function based on file extension
\itemize{
\item \code{.csv}: \link[readr:read_delim]{readr::read_csv}
\item \code{.json}: \link[jsonlite:read_json]{jsonlite::read_json}
\item \code{.rds}: \link[base:readRDS]{base::readRDS}
\item \code{.sav}: \link[haven:read_spss]{haven::read_sav}
\item \code{.xls}: \link{cloud_read_excel}
\item \code{.xlsx}: \link{cloud_read_excel}
\item \code{.xml}: \link[xml2:read_xml]{xml2::read_xml}
}
}

\examples{
\dontrun{
cloud_s3_read("data/mtcars.csv")
cloud_s3_read("models/random_forest.rds")
cloud_s3_read("data/dm.sas7bdat", fun = haven::read_sas)
}

}
