% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s3_transfer.R
\name{cloud_s3_upload}
\alias{cloud_s3_upload}
\title{Upload a local file to S3}
\usage{
cloud_s3_upload(file, project = getwd())
}
\arguments{
\item{file}{Path to a file relative to project folder root. Can contain only
letters, digits, '-', '_', '.', spaces and '/' symbols.}

\item{project}{Path to a project. By default it is current working directory.}
}
\description{
Uploads a file from project's folder to the corresponding
location on project's S3 folder
}
\examples{
\dontrun{
# uploads data/demo.csv to 'data' subfolder of project's S3 folder
cloud_s3_upload("data/demo.csv")
}

}
