% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cloudfs-package}
\alias{cloudfs}
\alias{cloudfs-package}
\title{cloudfs: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stefan Musch \email{stefan@gradientmetrics.com}

Authors:
\itemize{
  \item Tom Vladeck \email{tom@gradientmetrics.com }
  \item Iaroslav Domin \email{iaroslav@gradientmetrics.com}
  \item Michal Czyz \email{michal@gradientmetrics.com}
  \item Emmanuel Ugochukwu \email{emmanuel@gradientmetrics.com}
}

Other contributors:
\itemize{
  \item Gradient Metrics [copyright holder]
}

}
\keyword{internal}
