% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{cloudfs-package}
\alias{cloudfs}
\alias{cloudfs-package}
\title{cloudfs: Streamlined interface to interact with cloud storage platforms}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This package offers a unified interface for simplifying cloud storage interactions, including uploading, downloading, reading, and writing files, with functions for both Google Drive and Amazon S3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://g6t.github.io/cloudfs/}
  \item \url{https://github.com/g6t/cloudfs}
}

}
\author{
\strong{Maintainer}: Iaroslav Domin \email{iaroslav@gradientmetrics.com}

Authors:
\itemize{
  \item Stefan Musch \email{stefan@gradientmetrics.com}
  \item Michal Czyz \email{michal@gradientmetrics.com}
  \item Emmanuel Ugochukwu \email{emmanuel@gradientmetrics.com}
}

Other contributors:
\itemize{
  \item Gradient Metrics [copyright holder, funder]
}

}
\keyword{internal}
